+++ CHANGELOG.md
@@ -4,15 +4,25 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## [0.4.1.0] - 2023-XX-XX
+New release of the HyperDbg Debugger.
+
+### Added
+- **!crwrite** - Control Register Modification Event ([link](https://docs.hyperdbg.org/commands/extension-commands/crwrite))
+
+### Changed
+
 ## [0.4.0.0] - 2023-XX-XX
 New release of the HyperDbg Debugger.
 
 ### Added
 - The **!monitor** command now supports 'execution' interception ([link](https://docs.hyperdbg.org/commands/extension-commands/monitor))
 - **!crwrite** - Control Register Modification Event ([link](https://docs.hyperdbg.org/commands/extension-commands/crwrite))
+- **.pagein** - command is added to the debugger to bring pages in ([link](ttps://docs.hyperdbg.org/commands/meta-commands/.pagein))
 
 ### Changed
-- The '.start' command's mechanism for finding the entrypoint is changed to address issues ([link](https://github.com/HyperDbg/HyperDbg/commit/c64cb39ebe066865a85e9dd4b696295260a81f18))
+- The '.start' command's mechanism for finding the entrypoint is changed to address issues ([link](https://docs.hyperdbg.org/commands/meta-commands/.start))
 - The buffer overlap error in hyperlog in multi-core systems is fixed ([link](https://github.com/HyperDbg/HyperDbg/commit/1fa06c0b5a8b93656803fdc455025f59aadd0adb))
 - The implementation of 'dd' (define dwrod, 32-bit), and 'dw' (define word, 16-bit) is changed ([link](https://docs.hyperdbg.org/commands/scripting-language/assumptions-and-evaluations#keywords))
 - The problem with unloading driver (#238) is fixed ([link](https://github.com/HyperDbg/HyperDbg/issues/238))